fixed some minor bugs in labelbehaviorssocial.m set get_readframe_fcn interface with videoIO so that it does not use precise frames.

+++ trunk/matlab/filehandling/videoioreadframe.m
@@ -1,4 +1,4 @@
 function im = videoioreadframe(readerobj,f)
 
-seek(readerobj,f);
+seek(readerobj,double(f));
 im = getframe(readerobj);
